Move around code to expression propagation common
